@`README.md`; preview versus trunk branch use.
	@`.ssh/README.md`, @`posts/README.md`, @`hooks/README.md`; ditto.

Warning: This `preview` branch is for [beta tests (public review)](README.md#beta-testexperimental-builds) / [continuous integration (autonomous review)](https://google.com?q=continuous-integration-branch).
This is the first commit (limited to this `preview` branch, which is the branch used to backup&share local (draft) commits.)
The commits in this branch (commits which follow this commit): are not official, are not supported, are subject to frequent use of `git rebase` + `git push --force`
The commits before this commit exist (or soon will) in the `trunk` branch, which must pass all unit tests (autonomous review).
Once a commmit "Q" in the `preview` branch is fit for use:
	"Q" is moved (`git rebase -i HEAD~N`, where "N" is the count of the list of commits to go back through) before this commit,
	[this commit is now the ancestor of "Q", both now have new hashes]:
		`git switch trunk`,
		`git rebase -i HEAD`, `;%s/noop/pick Q/` #`git cherry-pick Q` changes the commit hash
		`git push origin HEAD:trunk` to the `trunk` branch for all to use.
	or (without switch from `preview`):
		`git push origin Q:trunk` #commits `Q` to the `trunk` branch (for all to use.)

Thus, this `preview` branch and the `trunk` branch will have a new common ancestor (which allows the 2 branches to continue as close as possible).